chore: remove ot definitions
