--- src/graphqlMappings.js
@@ -1,54 +1,0 @@
-import { definitions } from './openingHours';
-
-export default {
-  looProps: {
-    canHumanize: ['attended', 'babyChange', 'automatic', 'radar'],
-    definitions: {
-      opening: definitions,
-    },
-  },
-  humanizePropertyValue(val, property) {
-    if (this.looProps.definitions[property]) {
-      // We may use a human readable definition of this property value
-      let override = this.looProps.definitions[property].find(
-        (s) => s.value === val
-      );
-      if (override) {
-        return override.name;
-      }
-    }
-
-    if (this.looProps.canHumanize.includes(property)) {
-      // We can humanize this kind of property to make it more human-readable
-      return this.humanizeAPIValue(val, property);
-    }
-
-    // This was likely entered as human-readable, leave it be
-    return val;
-  },
-  humanizeAPIValue(val, property) {
-    // Unknown
-    if (val === undefined || val === null) {
-      return 'Not known';
-    }
-
-    // Capitalise
-    if (typeof val === 'string') {
-      val = val
-        .replace(/([A-Z])/g, ' $1')
-        .replace(/^./, (str) => str.toUpperCase())
-        .trim();
-    }
-
-    // Readable boolean
-    if (
-      typeof val === 'boolean' ||
-      (typeof val === 'string' &&
-        ['true', 'false'].indexOf(val.toLowerCase()) !== -1)
-    ) {
-      val = JSON.parse((val + '').toLowerCase()) ? 'Yes' : 'No';
-    }
-
-    return val;
-  },
-};
