+++ src/pages/AddPage.js
@@ -15,7 +15,6 @@ import useMapPosition from '../components/useMapPosition';
 import useNearbyLoos from '../components/useNearbyLoos';
 
 import config from '../config';
-import graphqlMappings from '../graphqlMappings';
 
 const UPDATE_LOO = loader('./updateLoo.graphql');
 
@@ -25,8 +24,6 @@ const initialFormState = {
 };
 
 const AddPage = (props) => {
-  const optionsMap = graphqlMappings.looProps.definitions;
-
   const [mapPosition, setMapPosition] = useMapPosition();
 
   const { data } = useNearbyLoos({
@@ -100,7 +97,6 @@ const AddPage = (props) => {
         saveLoading={saveLoading}
         saveResponse={saveResponse}
         saveError={saveError}
-        optionsMap={optionsMap}
         onSubmit={save}
       >
         <Box display="flex" flexDirection="column" alignItems="center">
